+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.3.5] - 2022-10-07
+### Added
+- Added extra logging. Logging Level can be set from config file.
+- Deprecated options MISCALE_VERSION, TIME_INTERVAL
+
 ## [0.3.4] - 2022-10-05
 ### Changed
 - Restored MQTT Discovery. ([fixes #65](https://github.com/lolouk44/hassio-addons/issues/65))